+++ docs/changelog.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- No json string message parsing in `sc-server` ([190](https://github.com/ostis-ai/sc-machine/issues/190))
 - `SCs-parser` create structure twice
 - `SCs-parser` doesn't unite all generated elements into structure ([205](https://github.com/ostis-ai/sc-machine/issues/205))
 - Link deletion ([193](https://github.com/ostis-ai/sc-machine/issues/193)